+++ elastic-job-lite/elastic-job-lite-spring/src/main/java/com/dangdang/ddframe/job/lite/spring/namespace/parser/script/ScriptJobBeanDefinitionParser.java
@@ -17,10 +17,13 @@
 
 package com.dangdang.ddframe.job.lite.spring.namespace.parser.script;
 
-import com.dangdang.ddframe.job.lite.spring.namespace.constants.ScriptJobBeanDefinitionParserTag;
+import com.dangdang.ddframe.job.config.script.ScriptJobConfiguration;
 import com.dangdang.ddframe.job.lite.spring.namespace.parser.common.AbstractJobBeanDefinitionParser;
+import org.springframework.beans.factory.config.BeanDefinition;
 import org.springframework.beans.factory.support.BeanDefinitionBuilder;
 import org.w3c.dom.Element;
+
+import static com.dangdang.ddframe.job.lite.spring.namespace.parser.script.ScriptJobBeanDefinitionParserTag.SCRIPT_COMMAND_LINE_ATTRIBUTE;
 
 /**
  * 脚本作业的命名空间解析器.
@@ -30,12 +33,10 @@ import org.w3c.dom.Element;
 public final class ScriptJobBeanDefinitionParser extends AbstractJobBeanDefinitionParser {
     
     @Override
-    protected Class<ScriptJobConfigurationDto> getJobConfigurationDTO() {
-        return ScriptJobConfigurationDto.class;
-    }
-    
-    @Override
-    protected void setPropertiesValue(final Element element, final BeanDefinitionBuilder factory) {
-        factory.addConstructorArgValue(element.getAttribute(ScriptJobBeanDefinitionParserTag.SCRIPT_COMMAND_LINE_ATTRIBUTE));
+    protected BeanDefinition getJobTypeConfigurationBeanDefinition(final BeanDefinition jobCoreConfigurationBeanDefinition, final Element element) {
+        BeanDefinitionBuilder result = BeanDefinitionBuilder.rootBeanDefinition(ScriptJobConfiguration.class);
+        result.addConstructorArgValue(jobCoreConfigurationBeanDefinition);
+        result.addConstructorArgValue(element.getAttribute(SCRIPT_COMMAND_LINE_ATTRIBUTE));
+        return result.getBeanDefinition();
     }
 }